src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filetype==1.2.0
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
Pillow==9.4.0
python-docx==0.8.11
requests==2.28.2
snowballstemmer==2.2.0
striprtf==0.0.22
textwrap3==0.9.2
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
